+++ psi4/src/psi4/libdpd/file2_mat_rd.cc
@@ -3,7 +3,7 @@
  *
  * Psi4: an open-source quantum chemistry software package
  *
- * Copyright (c) 2007-2018 The Psi4 Developers.
+ * Copyright (c) 2007-2019 The Psi4 Developers.
  *
  * The copyrights for code used from other parties are included in
  * the corresponding files.
@@ -52,7 +52,7 @@ int DPD::file2_mat_rd(dpdfile2 *File) {
         coltot = File->params->coltot[h ^ my_irrep];
 
         if (rowtot && coltot)
-            psio_read(File->filenum, File->label, (char *)File->matrix[h][0], rowtot * coltot * sizeof(double),
+            psio_read(File->filenum, File->label, (char *)File->matrix[h][0], static_cast<size_t>(rowtot) * coltot * sizeof(double),
                       irrep_ptr, &next_address);
     }
