+++ psi4/src/psi4/libsapt_solver/elst12.cc
@@ -3,7 +3,7 @@
  *
  * Psi4: an open-source quantum chemistry software package
  *
- * Copyright (c) 2007-2018 The Psi4 Developers.
+ * Copyright (c) 2007-2019 The Psi4 Developers.
  *
  * The copyrights for code used from other parties are included in
  * the corresponding files.
@@ -58,7 +58,7 @@ void SAPT2::elst12() {
 }
 
 double SAPT2::elst120(double **wBAA, double **wBRR, double **CHFA, int ampfile, const char *pAAlabel,
-                      const char *pRRlabel, const char *Ylabel, int foccA, int noccA, int nvirA) {
+                      const char *pRRlabel, const char *Ylabel, size_t foccA, size_t noccA, size_t nvirA) {
     int aoccA = noccA - foccA;
 
     double **pAA = block_matrix(aoccA, aoccA);
